\name{hDFPmcmc-package}
\alias{hDFPmcmc-package}
\alias{hDFPmcmc}
\docType{package}
\title{
\packageTitle{hDFPmcmc}
}
\description{
\packageDescription{hDFPmcmc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hDFPmcmc}
\packageIndices{hDFPmcmc}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{hDFPmcmc}

Maintainer: \packageMaintainer{hDFPmcmc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
